+++ caffe/cppbuild.sh
@@ -43,18 +43,18 @@ esac
 
 GLOG=0.3.4
 GFLAGS=2.1.2
-PROTO=2.6.1
-LEVELDB=1.18
+PROTO=3.1.0
+LEVELDB=1.19
 SNAPPY=1.1.3
 LMDB=0.9.18
-BOOST=1_61_0
+BOOST=1_62_0
 HDF5=1.8.17
-OPENBLAS=0.2.18
+OPENBLAS=0.2.19
 CAFFE_VERSION=master
 
 download https://github.com/google/glog/archive/v$GLOG.tar.gz glog-$GLOG.tar.gz
 download https://github.com/gflags/gflags/archive/v$GFLAGS.tar.gz gflags-$GFLAGS.tar.gz
-download https://github.com/google/protobuf/releases/download/v$PROTO/protobuf-$PROTO.tar.gz protobuf-$PROTO.tar.gz
+download https://github.com/google/protobuf/releases/download/v$PROTO/protobuf-cpp-$PROTO.tar.gz protobuf-$PROTO.tar.gz
 download https://github.com/google/leveldb/archive/v$LEVELDB.tar.gz leveldb-$LEVELDB.tar.gz
 download https://github.com/google/snappy/releases/download/$SNAPPY/snappy-$SNAPPY.tar.gz snappy-$SNAPPY.tar.gz
 download https://github.com/LMDB/lmdb/archive/LMDB_$LMDB.tar.gz lmdb-LMDB_$LMDB.tar.gz
@@ -107,7 +107,7 @@ cd ..
 
 cd leveldb-$LEVELDB
 make -j $MAKEJ
-cp -a libleveldb.a "$INSTALL_PATH/lib"
+cp -a out-static/libleveldb.a "$INSTALL_PATH/lib"
 cp -a include/leveldb "$INSTALL_PATH/include/"
 cd ..