+++ CHANGELOG.md
@@ -12,7 +12,7 @@
  * Make Caffe work on CPU-only machines ([issue #219](https://github.com/bytedeco/javacpp-presets/issues/219))
  * Fix loading issue with `opencv_face` ([issue bytedeco/javacv#470](https://github.com/bytedeco/javacv/issues/470))
  * Fix presets for CUDA on the `linux-ppc64le` platform
- * Upgrade presets for FFmpeg 3.1.2, x265 2.0, libvpx 1.6.0, CUDA 8.0, cuDNN 5.1, Caffe, TensorFlow 0.10.0
+ * Upgrade presets for FFmpeg 3.1.4, FFTW 3.3.5, GSL 2.2.1, LLVM 3.9.0, CUDA 8.0, cuDNN 5.1, Caffe, MXNet, TensorFlow 0.10.0, and their dependencies
  * Set default options in `tensorflow/cppbuild.sh` to prevent console reads during build
  * Add `Tensor.createStringArray()` method to access `DT_STRING` data ([issue #249](https://github.com/bytedeco/javacpp-presets/issues/249))
  * Fix Javadoc links for externally referenced classes